+++ kernel/src/drivers/kernel_logger.h
@@ -11,4 +11,4 @@ int printk(const char *fmt, ...);
 int serial_fprintk(const char *fmt, ...);
 int sprintf(char *buf, const char *fmt, ...);
 
-int sys_syslog(int type, const char *buf, size_t len);
+uint64_t sys_syslog(int type, const char *buf, size_t len);
